requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.1
celery[redis]>=5.3
redis>=5.0
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
django-cors-headers>=4.3
